refactor: move localstorage mock to a separate file
