+++ src/components/Widget/test/metadataInput.test.js
@@ -4,31 +4,10 @@ import { Provider } from 'react-redux';
 
 import Sender from '../components/Conversation/components/Sender';
 import { store, initStore } from '../../../store/store';
+import LocalStorageMock from './localStorageMock';
 
 
 describe('Metadata store affect input behavior', () => {
-  class LocalStorageMock {
-    constructor() {
-      this.store = {};
-    }
-
-    clear() {
-      this.store = {};
-    }
-
-    getItem(key) {
-      return this.store[key] || null;
-    }
-
-    setItem(key, value) {
-      this.store[key] = value.toString();
-    }
-
-    removeItem(key) {
-      delete this.store[key];
-    }
-  }
-
   const localStorage = new LocalStorageMock();
   const stubSocket = jest.fn();
   initStore('dummy', 'dummy', stubSocket, localStorage);
